requirements.txt
openai
chroma
langchain
python-dotenv
chromadb
tiktoken
